Adds `ImperativeIterator` utility.
This is an implementation of an `AsyncIterator` which is more ergonomic for imperative use cases. It provides `emit`, `error`, and `done` methods for producing data in an iterator while still implementing the `AsyncIterator` contract. This makes it easier to produce iterators in contexts where real generators can be hard to work with due to scoping issues.