Adds `hiddenChild` property.
This allows routes to be hidden from the forest passed into the `render` function of top-level routes. This is very useful when this forest is used to render links in site navigation, as hidden links will naturally be removed from this visual list. Hidden routes are still rendered and accessible to any deep link to the right path, they are only omitted from the object passed into the `render` function.

While this is very useful for `@rules_prerender` docs, I'm skeptical that this will be a good idea for a long-term officially supported routing API for a few reasons.
1.  `hiddenChild` is a weird name but the best I could come up with to express that the route is remove from the `children` array of its parent, yet the child still has a valid reference to its parent.
2.  Hiding parent routes is a bit awkward, as it sections off that subtree, but everything in that subtree still obeys the same internal rules. Meaning you can have non-hidden routes under a hidden parent. These routes can find each other through `parent` links, but are not accessible from the root.
3.  It's potentially confusing that hidden routes are still rendered and accessible to user. They are only removed from the data structure used to generate site navigation.
4.  The value of this feature is hard to intuit without being directly told. This requires a developer to go through the process:
    *   "`hiddenChild` hides from the input of `render` but still renders the route."
    *   "The input route forest of `render` is very useful to render site navigation."
    *   "Therefore, this is a useful feature to hide specific routes from site navigation but still render and have individual links to them."

A potential alternative approach is to make the route config generic with some user-specified `data` property. Then users could define their own `hidden` property with its own semantics which can be built into their navigation rendering. This would be a more obvious primitive to build on top of, but as this commit demonstrates getting the behavior right is actually a lot more nuanced than it initially seems.