Fix text wrapping when animating the nav pane.
When animating the link text would start wrapping as the nav pane shrunk, which looked very awkward. This introduces a container element which is shrunk while the `rp-nav-pane` element remains fixed at the same size. This prevents any text underneath it from wrapping.

At the same time, I also updated the nav pane to translate left at the same pace as the width is shrunk, so this looks like the nav pane "moving left" as opposed to "covered from the right".