Hides nav pane by default for small viewports.
This is easy to do with media queries, but required changes to the JS and CSS to manage the toggle. We can't use `.nav-collapsed` anymore because the absence of this class means the default behavior must be used, which could be now be expanded or collapsed, based on the media query. Instead, I switched to a trinary/enum state of `nav="collapsed"`, `nav="expanded"`, and `:not([nav])`. When no `nav` is provided, the default value is used based on a media query of viewport width. On component hydration, it makes the same media query and initializes the `nav` attribute. From there it can toggle as expected.

Not a huge fan of this approach, but I think it is the best option I can see right now.

https://techhub.social/@develwithoutacause/111432914215810488