Enables `isolatedModules`.
This allows for TS compilations to be parallelized by ensuring that the emit for each file does not depend on types from other files. I'm not really taking advantage of this parallelism right now, but it's a good constraint to maintain so we can take advantage of it in the future.

Supporting it really isn't that difficult. The main issue which came up here is that exporting a type now requires an explicit `export type` so TS is aware that the imported value is a type without having to resolve the dependency.