Use the `workspace:` protocol for local dependencies.
I discovered that we were installing `rules_prerender@0.0.0-releasetest`, which is clearly incorrect as we have a local build of `rules_prerender` to use instead. The best fix I found was to use pnpm's `workspace:` protocol to link internal dependencies. This removed the dependency on the `0.0.0-releasetest` version.

Unfortunately the `npm_package` rule does not seem to strip `workspace:` protocols for published packages, so I needed to do this manually in our custom stamping logic. I also updated the external example to explicitly depend on the `rules_prerender` packages and be linked into the pnpm workspace. This revealed a strict dependency error on `@types/node`. Apparently `rules_prerender` includes `Buffer` which comes from `@types/node`, and this bleeds into the `*.d.mts`, so downstream consumers do need that type. Therefore I had to move it from `devDependencies` to `dependencies` to make sure it's available.