Bump Node.js version to `18.17.0`.
This appears to be the latest Node version supported by `@aspect_rules_js` v1 at the moment due to an older dependency on `@rules_nodejs` which can't be bumped locally (it's not a `_maybe` "peer" dep).

* https://github.com/aspect-build/rules_js/blob/c0d78c04141ba341565107d130b04a8504c4ff3b/js/repositories.bzl#L17-L22
* https://github.com/bazelbuild/rules_nodejs/blob/c0b5865d7926298206e9a6aa32138967403ad1f2/nodejs/private/node_versions.bzl#L2306-L2313

This also implicit bumps the `pnpm` version. I also updated the `engines` field and dropped `npm` since that's not supported package manager for this repo.