Updates ESLint configuration to enable `strict-type-checked` and `stylistic-type-checked` rule sets and fix errors.
These enable a whole bunch of checks by default which rely on the type checker. This slows down the process a bit, but for a project of this size it's manageable. I fixed all the issues they reported and reviewed how much I liked the reports it gave. I disabled and reconfigured a few checks to align with my (correct) sensibilities.