Formats all Starlark files with Buildifier.
https://github.com/bazelbuild/buildtools/tree/main/buildifier

This is limited to Starlark for now since that's likely where it's most important to stay consistent. Buildifier should be run as part of a CI check to ensure things stay formatted, but I don't have a Go dependency right now and don't want to add one just for this. When we inevitably add one for Gazelle, we can reconsider setting up Buildifier directly in the repo.