Switches from `types_only` to `prune_dependencies` as a mechanism for removing workspace-local package dependencies.
`types_only` breaks in `@aspect_rules_js` v1.40.1 because declarations are dropped from `npm_link_package_store` in https://github.com/aspect-build/rules_js/pull/1554, therefore we need a different approach.

Unfortunately there does not appear to be a clean way to depend on only the types of a module, so instead I added `prune_dependencies` which attempts to remove type-only dependencies after the fact. This allows `//packages/preact/...` to depend on the `rules_prerender` NPM package locally in the workspace, but then have that dependency removed. When the package is installed and linked in the user's workspace, `@user//:node_modules/@rules_prerender/preact` is linked against the user's `@user//:node_modules/rules_prerender` instead of `@rules_prerender//:node_modules/rules_prerender`.

`//tools/binaries/renderer` usage of `types_only` is not strictly necessary as it (intentionally) does have a value import of `rules_prerender`. It does technically depend on the local copy of `@rules_prerender//:node_modules/rules_prerender`, but this is never executed at runtime so it's not a significant problem. As long as we limit it to type-only imports, it should be fine.

`prune_dependencies` only works on an `npm_package`, meaning we can't use it the `:renderer` `ts_project`. I did consider adding another version of `prune_dependencies` which prunes `JsInfo` providers instead, however this seems to be a more complicated structure with file paths and nested providers which need to be pruned. Since it's not strictly necessary for `//tools/binaries/renderer`, I opted not to go down that route just yet, but it can be considered in the future.

I also briefy attempted creating a `@rules_prerender/renderer` package just to apply `prune_dependencies` to that. However `:renderer` depends on content outside its directory such as `//common:formatters`, so this needs to be included in the package as well, and an `npm_package` in `//tools/binaries/renderer/...` cannot include content outside of it. I then tried moving the `npm_package` to `//:renderer_pkg`, but this requires a `package.json` in the same directory, which would conflict with the `rules_prerender` package which has the same constraint. So there's no easy way to hack around this. Instead, we'll just rely on avoiding a value import of `rules_prerender` in the renderer.