Bumps `netlify-cli` to the current latest version (17.33.5).
This actually broke deployments because of [this commit](https://github.com/netlify/cli/commit/f684887c6041cbad4d4cc2ebabb70fb2bbf741b4) which seems to resolve things based on the CWD instead of the "root" (resolves to `.../execroot/rules_prerender`). We actually would probably prefer the CWD here since that would allow `rootpath` over `execroot` which is better because `@aspect_rules_js` tries copies everything into a single directory.

However the `netlify.toml` config file is not picked up under `--dir` and there does not appear to be an option to provide it explicitly. The easier approach is to `chdir` into the root-relative path (under runfiles) and deploy from there.

Unfortunately the transitive `unix-dgram` dependency does not appear to have been removed or updated, so the pnpm package extension still seems to be required.