Add `--dir .` to Netlify deployment.
Bazel 7 appears to have change something such that `netlify-cli` interprets its environment slightly differently, resulting in it inferring a `dist/` path which does not existing in the output (`dist/` is used as a symlink, but does not appear in the actual sandbox). By passing `--dir .`, we force Netlify to deploy the CWD, rather than attempting to infer another directory.