Updates comment for `--incompatible_disable_native_repo_rules`.
Tried to enable this and found one missing `load` statement in the external example. However this option appears to just be broken with `--enable_bzlmod` as it triggers an error in a non-existant `WORKSPACE` file, even if the actual `WORKSPACE` file is empty. I filed https://github.com/bazelbuild/bazel/issues/23206, as it appears to be a core Bazel bug. `--enable_bzlmod` breaks a bunch of things here and I haven't had time to properly look into it, so I think it's best to just leave `--incompatible_disable_native_repo_rules` off for now.