Update `ts_project` to automatically include a `package.json` file.
The `ts_project.bzl` file contains some comments with more details, but TL;DR: we need an explicit dependency on a `package.json` file so TypeScript can find `type: "module"` for `*.tsx` files and compile them as ESM. This is necessary to be compatible with Node 16 module resolution. Adding these dependencies would be quite infeasible to do manually, so we encode the convention into `ts_project` for convenience.