WIP: Bump to Node `18.20.3`.
We need Node 18.20.0+ for module hooks, so I'm bumping to the latest available Node 18 release (just avoiding breaking changes for the moment). The latest release is technically 18.20.5, but 18.20.3 is the latest `@rules_nodejs` v5 will go. Unfortunately the latest `@aspect_rules_js` v1 (1.42.3) depends on a slightly out of date `@rules_nodejs` (5.8.4), so I had to manually override that version to the actual latest v5 (5.8.5).

See:
* https://github.com/bazel-contrib/rules_nodejs/blob/5.8.5/nodejs/private/node_versions.bzl#L2386-L2393
* https://github.com/aspect-build/rules_js/blob/v1.42.3/js/repositories.bzl#L17-L22