WIP: Adds `key` to inline scripts.
This allows the bundler to disambiguate and ignore a single inline script in a single component which is rendered multiple times.

At rendering time, we can't know that either returned `VNode` will actually end up in the final HTML page (they could be dropped at a later step), therefore we have to produce the entire inline script. The key value allows annotation extractor to deduplicate same-key scripts to avoid bundling multiple copies.

`key` itself is a tagged template literal which uses the identity of the literal array to cache and reuse a key name, meaning devs don't need to namespace them.

```tsx
<>
  {/* Different keys! Both are logged. */}
  {inlineScript(key`log`, `console.log('first');`)}
  {inlineScript(key`log`, `console.log('second');`)}
</>
```

However, reusing the same key ensures only the first script is used, and since the tagged template literal marks the call site a very simple refactoring like:

```tsx
<>
  {log(`console.log('first');`)}
  {log(`console.log('second');`)}
</>

function log(script: string): VNode {
  // Same key! Only `first` is logged.
  return inlineScript(key`log`, script);
}
```

Actually has a different output. Because of this confusing behavior, I'm honestly not sure this is a good idea. It's helpful to avoid the need for globally unique keys, but can lead to very unintuitive bugs.